Kernel: Move VirtIO console code to the Devices/Serial/VirtIO directory
