+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -495,8 +495,6 @@ executable("Kernel_bin") {
     "Bus/USB/USBManagement.cpp",
     "Bus/USB/USBPipe.cpp",
     "Bus/USB/USBTransfer.cpp",
-    "Bus/VirtIO/Console.cpp",
-    "Bus/VirtIO/ConsolePort.cpp",
     "Bus/VirtIO/Device.cpp",
     "Bus/VirtIO/Queue.cpp",
     "Bus/VirtIO/RNG.cpp",
@@ -553,7 +551,8 @@ executable("Kernel_bin") {
     "Devices/HID/PS2/MouseDevice.cpp",
     "Devices/KCOVDevice.cpp",
     "Devices/KCOVInstance.cpp",
-    "Devices/PCISerialDevice.cpp",
+    "Devices/Serial/VirtIO/Console.cpp",
+    "Devices/Serial/VirtIO/ConsolePort.cpp",
     "Devices/SerialDevice.cpp",
     "Devices/Storage/ATA/AHCI/Controller.cpp",
     "Devices/Storage/ATA/AHCI/InterruptHandler.cpp",
